fixing import errors; moved files/folders; fixed my mistake of installing redux packages on server, so it is now on client
