+++ client/package.json
@@ -11,6 +11,9 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-redux": "^7.2.4",
+    "redux-form": "^8.3.7",
+    "redux-thunk": "^2.3.0",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
