+++ client/src/index.js
@@ -1,4 +1,4 @@
-import "./index.css";
+import "./index.scss";
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
