+++ client/src/index.scss
@@ -10,7 +10,7 @@
 html {
   box-sizing: border-box;
   font-size: 1rem;
-  
+  font-family: "Commissioner", sans-serif;
 }
 a {
   text-decoration: none;
