+++ client/src/pages/ErrorPage/ErrorPage.js
@@ -1,11 +1,8 @@
-
-
-import "./ErrorPage.scss";
-
 import React from "react";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
-import history from "../../../history";
+import history from "../../history";
+import "./ErrorPage.scss";
 
 class ErrorPage extends React.Component {
   state = {};
