+++ index.js
@@ -7,7 +7,7 @@ const cors = require("cors");
 const helmet = require("helmet");
 const noCache = require("nocache");
 const logger = require("morgan");
-const loginRouter = require("../controllers/login");
+const loginRouter = require("./server/controllers/login");
 
 const app = express();
 app.use(cors());
