+++ package.json
@@ -33,14 +33,12 @@
     "dotenv": "^9.0.1",
     "express": "^4.17.1",
     "helmet": "^4.6.0",
+    "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.6.6",
     "mongoose": "^5.12.7",
     "morgan": "^1.10.0",
     "nocache": "^2.1.0",
-    "nodemon": "^2.0.7",
-    "react-redux": "^7.2.4",
-    "redux-form": "^8.3.7",
-    "redux-thunk": "^2.3.0"
+    "nodemon": "^2.0.7"
   },
   "devDependencies": {
     "prettier": "^2.3.0"
